pyproject.toml
[tool.poetry]
name = "livraria"
version = "0.1.0"
description = ""
authors = ["Flavio Carvalho <professorflaviosoares@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.3"
factory-boy = "^3.3.1"
django = "^5.1.2"
django-rest-framework = "^0.1.0"
django-extensions = "^3.2.3"
django-debug-toolbar = "^4.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
